requirements.txt
# add your project requirements here, be sure to pin the versions
pytest==7.2.2
pandas==2.2.3
numpy==2.1.1
boto3==1.35.30
python-dotenv==1.0.1
ipykernel==6.29.5